MADE IT WORK POG now i just need to do optimizations and stuff

+++ util.h
@@ -21,7 +21,7 @@ CharNode* charnode_insert(CharNode* pnode, vec* size);
 
 typedef struct gamestate {
   double delta; // Delta time in the last frame
-  float last;
+  double last;
   int width, height; // Starting width and height of window
 } gamestate;
